LTCGExtensions: remove `std::clamp` functions.
C++17 introduced a new `std::clamp` function that behaves differently
than the one implemented in LTCGExtensions: it doesn't accept inverted
range (lo > hi), and the result is undefined. Since this means that our
(illegal) extension to STL is now breaking the definition of this
function, we have to remove declarations of it.

Practically, calls to `std::clamp` will still work for `CGFloat`s and
types that implement the `<` operator. However, it is no longer legal to
pass an inverted range to `std::clamp`, so this is a breaking API
change.

Finally, we implement the `CGPoint` clamping function to `CGRect` as a
regular LTCGExtension function named `CGPointClampToRect`, since the
behavior there is completely different than what expected in
`std::clamp`.

+++ Pinky/Pinky/PNKImageMotionTreeLayer.mm
@@ -116,7 +116,7 @@ NS_ASSUME_NONNULL_BEGIN
       auto xRelativeDisplacement = (half_float::half)(-xDisplacement * inverseWidth);
       auto yRelativeDisplacement = (half_float::half)(-yDisplacement * inverseHeight);
 
-      int newRow = std::clamp(row + std::round(yDisplacement), 0, _treesWithLabels.rows - 1);
+      int newRow = std::clamp(row + int(std::round(yDisplacement)), 0, _treesWithLabels.rows - 1);
       auto pointerToNewRow = (half_float::half *)displacements->ptr(newRow);
 
       for (int col = left; col < left + width; ++col) {
@@ -124,7 +124,7 @@ NS_ASSUME_NONNULL_BEGIN
           continue;
         }
 
-        int newCol = std::clamp(col + std::round(xDisplacement), 0, _treesWithLabels.cols - 1);
+        int newCol = std::clamp(col + int(std::round(xDisplacement)), 0, _treesWithLabels.cols - 1);
 
         pointerToNewRow[2 * newCol] = xRelativeDisplacement;
         pointerToNewRow[2 * newCol + 1] = yRelativeDisplacement;
